prototype/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
pandas
joblib
scikit-learn
openai
python-dotenv
google-generativeai
